app/requirements.txt
wait-for-it
Django>=3.0,<4.0
psycopg2>=2.8
djangorestframework